docs: improve examples
