+++ ecdsa_test.go
@@ -1,11 +1,7 @@
 package ekliptic
 
 import (
-	cryptorand "crypto/rand"
-	"crypto/sha256"
-	"fmt"
 	"math/big"
-	mathrand "math/rand"
 	"testing"
 
 	"github.com/kklash/ekliptic/test_vectors"
@@ -45,30 +41,4 @@ func BenchmarkSignECDSA(b *testing.B) {
 			r, s,
 		)
 	}
-}
-
-func ExampleSignECDSA() {
-	randReader := mathrand.New(mathrand.NewSource(1))
-
-	key, _ := cryptorand.Int(randReader, Secp256k1_CurveOrder)
-	nonce, _ := cryptorand.Int(randReader, Secp256k1_CurveOrder)
-
-	hashedMessage := sha256.Sum256([]byte("i love you"))
-	hashedMessageInt := new(big.Int).SetBytes(hashedMessage[:])
-
-	r := new(big.Int)
-	s := new(big.Int)
-
-	SignECDSA(
-		key, nonce, hashedMessageInt,
-		r, s,
-	)
-
-	fmt.Printf("r: %x\n", r)
-	fmt.Printf("s: %x\n", s)
-
-	// output:
-	//
-	// r: 4a821d5ec008712983929de448b8afb6c24e5a1b97367b9a65b6220d7f083fe3
-	// s: 2e4f380e0ea1dfcb7cced430437c98b4570a06b3e929a3b19e6bbd53df2cf3f6
 }
