+++ multiply_base_test.go
@@ -1,7 +1,6 @@
 package ekliptic
 
 import (
-	"fmt"
 	"math/big"
 	"testing"
 
@@ -41,22 +40,4 @@ func BenchmarkMultiplyBasePoint(b *testing.B) {
 	for i := 0; i < b.N; i++ {
 		MultiplyBasePoint(vector.K, x, y)
 	}
-}
-
-func ExampleMultiplyBasePoint() {
-	// Generate a public key from a private key.
-	privateKey, _ := new(big.Int).SetString("c370af8c091812ef7f6bfaffb494b1046fb25486c9873243b80826daef3ec583", 16)
-	x := new(big.Int)
-	y := new(big.Int)
-
-	MultiplyBasePoint(privateKey, x, y)
-
-	fmt.Println("Public key:")
-	fmt.Printf(" x: %x\n", x)
-	fmt.Printf(" y: %x\n", y)
-
-	// output:
-	// Public key:
-	//  x: 76cd66c6cca75278ff408ce67290537367719154ae2b96448327fe4033ddcfc7
-	//  y: 35663ecbb64397bb9bd79155a1e6b138c2fb8fa1f11355f8e9e97ddd88a78e49
 }
